% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asmntDashboard.r
\name{asmntDashboard}
\alias{asmntDashboard}
\title{Run the assessment dashboard application
Runs the assessment dashboard shiny app embedded in UT IR tools package.}
\usage{
asmntDashboard()
}
\description{
Run the assessment dashboard application
Runs the assessment dashboard shiny app embedded in UT IR tools package.
}
