% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessEcoli.r
\name{assessEColi}
\alias{assessEColi}
\title{Assess E.coli data at the year/site level.}
\usage{
assessEColi(data, rec_season = TRUE, SeasonStartDate = "05-01",
  SeasonEndDate = "10-31")
}
\arguments{
\item{data}{A prepped dataframe object (likely the ecoli object within the prepped_data list--prepped_data$ecoli) containing e.coli data at the site/use/parameter level with standards assigned.}

\item{rec_season}{Logical. If TRUE, restricts assessments to recreation season data.}

\item{SeasonStartDate}{A string in the form "mm-dd" to define beginning of rec season over which to perform assessments.}

\item{SeasonEndDate}{A string in the form "mm-dd" to define end of rec season over which to perform assessments.}
}
\value{
Returns list with two objects: assessments from all Scenarios on all data, and ecoli assessments aggregated over scenario and year and rolled up to site level.
}
\description{
Compares E.coli data to 30-day and max criterion standards using Scenarios A, B, and C, and assigns e. coli assessment categories to each site.
}
