% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessExcCounts.r
\name{assessExcCounts}
\alias{assessExcCounts}
\title{Perform assessments based on sample & exceedance counts (conventional & toxic assessments)}
\usage{
assessExcCounts(data, min_n, max_exc_count = NA, max_exc_pct = NA,
  max_exc_count_id = NA, max_exc_pct_id = NA, id_cols = c())
}
\arguments{
\item{data}{Site-use-parameter-criterion specific sample and  exceedance counts(i.e. output from countExceedances() )}

\item{min_n}{Minimum sample size for sufficient data assessment (sample counts >= min_n are considered sufficient)}

\item{max_exc_count}{Maximum allowable exceedance count for full support (exceedance counts > max_exc_count are considered not supporting) - one of max_exc_count or max_exc_pct must be specified}

\item{max_exc_pct}{Maximum allowable exceedance percentage for full support (exceedance pcts > max_exc_pct are considered not supporting - one of max_exc_count or max_exc_pct must be specified}

\item{max_exc_count_id}{Maximum allowable exceedance count for insufficient data with exceedances (for sites/parameters with insufficient data, exceedance counts > max_exc_count_id are considered insufficient data with exceedances) - one of max_exc_count_id or max_exc_pct_id must be specified}

\item{max_exc_pct_id}{Maximum allowable exceedance percentage for insufficient data with exceedances (for sites/parameters with insufficient data, exceedance pcts > max_exc_count_id are considered insufficient data with exceedances) - one of max_exc_count_id or max_exc_pct_id must be specified}
}
\value{
Returns input data frame with site-use-parameter-criterion specific assessment categories appended.
}
\description{
Performs site level (site-parameter-use-criterion specific) water quality assessments based on sample & exceedance counts by applying IR method flowcharts with user specified inputs.
}
\details{
When percentage arguments are supplied, specified percentages are used to calculate the maximum allowable number of whole samples exceeding a standard for that site/parameter to be considered supporting. Note that the min_n argument is >= and all other arguments are >.
}
