% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessLakeProfiles.r
\name{assessLakeProfiles}
\alias{assessLakeProfiles}
\title{Run lake profile assessments}
\usage{
assessLakeProfiles(
  data,
  do_crit = list(`3A` = 4, `3B` = 3),
  temp_crit = list(`3A` = 20, `3B` = 27),
  uses_assessed = c("3A", "3B")
)
}
\arguments{
\item{data}{Lake profiles object returned by dataPrep step.}

\item{do_crit}{List of beneficial use classes and associated dissolved oxygen criteria to use for assessment. Defaults to list("3A"=5, "3B"=3). This excludes chronic & ELS present criteria per assessment methods. Sites with site specific criteria will be assessed regardless of criteria specified in this argument. Objects in this list should match the uses_assessed argument.}

\item{temp_crit}{List of beneficial use classes and associated water temperature criteria to use for assessment. Defaults to list("3A"=20, "3B"=27). This excludes chronic & ELS present criteria per assessment methods. Sites with site specific criteria will be assessed regardless of criteria specified in this argument. Objects in this list should match the uses_assessed argument.}

\item{uses_assessed}{Vector of beneficial uses to be assessed for lake profiles. Defaults to 3A & 3B uses.}
}
\value{
Returns a list of lake profile assessment dataframes. profile_asmnts_mlid_param contains site/parameter level profile assessments, profile_asmnts_individual contains assessments for each individual profile,
	profile_criteria contains the criteria used for the profile assessment (excluding any site-specific criteria that may have occurred in the input dataset), 
 profiles_long contains profile data in long format including the numeric criterion associated with each parameter, profiles_wide contains profile data cast to wide format.
}
\description{
Performs lake profile assessments per IR assessment methods. This includes identifying stratified profiles & applying appropriate assessment methods for stratified & non stratified profiles.
}
