% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCriteria.r
\name{assignCriteria}
\alias{assignCriteria}
\title{Assign numeric criteria}
\usage{
assignCriteria(data, translation_wb, sheetname, startRow = 1,
  browsefile = F)
}
\arguments{
\item{data}{A merged, parameter translated, spatially referenced WQP result object.}

\item{sheetname}{Name of sheet in workbook holding criteria to be assigned.}

\item{startRow}{Row to start reading excel sheet from (in case headers have been added). Defaults to 1.}

\item{crit_wb}{Full path and filename for workbook containing criteria.}
}
\value{
A flattened & expanded sample x use x parameter x criterion data.frame.
}
\description{
Assigns general & site-specific numeric WQ criteria to WQP data. Does not calculate criteria w/ correction factors. See calcCriteria (function in development).
}
