% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoValidateWQPsites.r
\name{autoValidateWQPsites}
\alias{autoValidateWQPsites}
\title{Auto-validate USEPA WQP stations}
\usage{
autoValidateWQPsites(sites_file, master_site_file, polygon_path, outfile_path,
  correct_longitude = TRUE,
  site_type_keep = c("Lake, Reservoir, Impoundment", "Stream",
  "Stream: Canal", "Stream: Ditch", "Spring", "River/Stream", "Lake",
  "River/Stream Intermittent", "River/Stream Perennial", "Reservoir",
  "Canal Transport", "Canal Drainage", "Canal Irrigation"))
}
\arguments{
\item{sites_file}{Full path and filename of sites file queried from WQP to be reviewed (.csv).}

\item{master_site_file}{Full path and filename of master site list as generated by autoValidateWQPsites function and manual site review application (.csv).}

\item{polygon_path}{Full path to folder containing, AU, land ownership, and Utah state boundary 1000 m buffer polygon shapefiles.}

\item{outfile_path}{Path for file outputs.}

\item{site_type_keep}{Vector of site type names to be considered in assessment process. Non-specified site types will be automatically rejected. Additional site types can be excluded when running assessment tools. However, including extra site types here may necessetitate additional manual site reviews.}

\item{ownership_assess}{Vector of property ownership labels to be considered in assessment process. Defaults to c("Federal","Private","State") and exclude "Tribal".}
}
\value{
Exports a new, undated master site list to the outfile_path. If one already exists in outfile_path, it is moved to the 'edit_logs' folder and renamed with the system date.
}
\description{
Performs auto-validation on previously queried WQP stations combined with the existing master site list.
Auto-validates any new sites (i.e. those not currently in the master site list) and appends them into existing master site list flagged for acceptance, rejection, or review.
Checks for any new site types in new data. A warning message and a list of new site types is printed if new site types are encountered.
Also re-auto-validates any sites in the master site list that have previously only undergone auto-validation (to account for any changes in auto validation process) and checks existing master site list for changes in AUs, selected site types, or property boundaries.
}
