% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{chunk}
\alias{chunk}
\title{Divide a dataframe column into vectors, each the length of a specified increment}
\usage{
chunk(.data, col, increment)
}
\arguments{
\item{.data}{The dataframe containing the column to be divided}

\item{col}{The column to be divided, quoted}

\item{increment}{The increment by which to divide the column}
}
\value{
Returns a list of vectors that together contain a non-duplicated set of all unique values in the column. Each vector is the length of the specified increment, except for a possible smaller vector that contains remaining values.
}
\description{
chunk() takes a dataframe column and divides it into vectors the length of a specified increment, plus a smaller vector if there are any remaining values. The vectors are saved as a list, which can be used to filter in a looped database query. The function was created as a means to filter by local values when the number of such values exceeds the filter limit.
}
