% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countExceedances.r
\name{countExceedances}
\alias{countExceedances}
\title{Calculate sample and exceedance counts (conventional & toxic assessments)}
\usage{
countExceedances(data, group_vars = c("IR_MLID", "R317Descrp", "IR_Lat",
  "IR_Long", "ASSESS_ID", "AU_NAME", "BeneficialUse", "R3172ParameterName",
  "CriterionLabel", "SSC_MLID", "AsmntAggFun"))
}
\arguments{
\item{data}{A prepped water quality portal data object (i.e. output from dataPrep() )}

\item{group_vars}{Vector of column names on which to group data when calculating sample counts and exceedances. This should not include any factors that prevent aggregation to site-scale assessments (e.g. date, time, etc.). See default for recommended.}
}
\value{
Returns sample and exceedance counts aggregated by grouping variables.
}
\description{
Compares water quality result values to standards to calculates sample and exceedance counts. This is geared towards conventional and toxic assessments.
}
