% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.r
\name{dataPrep}
\alias{dataPrep}
\title{Final data prep step before assessments}
\usage{
dataPrep(data)
}
\arguments{
\item{data}{A merged, translated, and numeric criteria assigned WQP results R-object. Target units for conversions are defined by units associated with assigned numeric critera.}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{sheetname}{Name of sheet in workbook holding IR unit conversion table. Defaults to "unitConvTable".}

\item{startRow}{Row to start reading the unit conversion table excel sheet from (in case headers have been added). Defaults to 1.}
}
\value{
WQP data input with .
}
\description{
Performs unit conversions & daily data aggregations. Also checks native vs. target activities and fractions, dissolved vs. total value checks, river/stream depth & flow checks, and generates value based data flags.
}
