% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Make Sherman Font available upon package launch}
\usage{
.onLoad(libname, pkgname)
}
\description{
theme_ir_fonts() is a wrapper function for showtext_auto() that makes Syracuse University Shermans Fonts available upon loading the iRtools package.
}
