% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel_out.R
\name{excel_out}
\alias{excel_out}
\title{Export data frames to Excel}
\usage{
excel_out(
  .data,
  path,
  output = "tab",
  tab.names = FALSE,
  title = NULL,
  subtitle = NULL,
  footnote = NULL
)
}
\arguments{
\item{.data}{A data frame or list of data frames wrapped in a list. Use list(df1, df2, etc.).}

\item{path}{A file path to export to.}

\item{output}{The desired output.  Use 'tab' to write each data frame to its own worksheet.  Use 'stack' to stack all data frames in one worksheet.  Defaults to 'tab'.}

\item{tab.names}{Logical.  Activated when output is set to 'tab'.  Tab names must be passed in the list of data frames for x (i.e. list(Name1 = df1, Name2 = df2, etc.)).  Defaults to FALSE.}

\item{title}{A single title or vector of titles.  If one title is listed, then there must be one title for each data frame.  Defaults to NULL.}

\item{subtitle}{A single subtitle or vector of subtitles.  If one subtitle is listed, then there must be one subtitle for each data frame.  Defaults to NULL.}

\item{footnote}{A single footnote or vector of footnotes.  If one footnote is listed, then there must be one footnote for each data frame.  Defaults to NULL.}
}
\description{
One or more data frames can be exported to Excel in one of two ways: one per worksheet or all stacked in a single worksheet
}
