% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMaskedValues.R
\name{fillMaskedValues}
\alias{fillMaskedValues}
\title{Select quantitation limits, determine detection condition, and fill masked values as appropriate.}
\usage{
fillMaskedValues(results, detquantlim, translation_wb,
  detsheetname = "detLimitTypeTable", unitsheetname = "unitConvTable",
  detstartRow = 3, unitstartRow = 1, unitstartCol = 1,
  lql_fac = 0.5, uql_fac = 1)
}
\arguments{
\item{results}{A WQP results (must include narrow result, merged, wide objects OK) R-object name.}

\item{detquantlim}{A WQP detection/quantitation limit file R-object. Should be matching complement to WQP results input.}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{detsheetname}{Name of sheet in workbook holding detection limit type names and ranked prioritizations table. Defaults to "detLimitTypeTable".}

\item{unitsheetname}{Name of sheet in workbook holding unit conversion table. Defaults to "unitConvTable".}

\item{detstartRow}{Row to start reading the detLimitTypeTable excel sheet from (in case headers have been added). Defaults to 3.}

\item{unitstartRow}{Row to start reading the unitConvTable excel sheet from (in case headers have been added). Defaults to 1.}

\item{unitstartCol}{Column to start reading the unitConvTable excel sheet from (in case headers have been added). Defaults to 1.}

\item{lql_fac}{Numeric - factor by which to multiply lower quantitation limit type values when filling masked data or other non-detects (e.g. below lql values). Default = 0.5.}

\item{uql_fac}{Numeric - factor by which to multiply upper quantitation limit type values when filling masked data or other over limit values. Default = 1.}
}
\value{
Returns a data frame with new columns of selected limits, filled values, units, and detection condition appended to results input. In IR_DetCond column, "ND"=non-detect, "OD"=over detection, "NRV"=no result value & no ranked detection limit, "DET"=detection.
}
\description{
This function selects a single upper and lower quantitation limit for each record in result (if one exists in detquantlim), determines detection condition for each record based on the presence/absence of a result value, upper or lower quant limits, and whether the result value is above or below quant limits or <=0.
}
