% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_connect.R
\name{ir_connect}
\alias{ir_connect}
\title{Connect to an Oracle database}
\usage{
ir_connect(.dsn, access_db = FALSE, out = "assign")
}
\arguments{
\item{.dsn}{Data Source Name (DSN) of a valid locally-installed ODBC driver (64-bit Oracle in OraClient 12Home1)}

\item{access_db}{Is .dsn a microsoft access database}

\item{out}{Type of output desired. The defalt, "assign", will return the connection as an object in the global environment with the same name as the DSN. Alternatively, "return" will simply return the connection. The "return" option should be used only when using oira_connect directly inside another function call such as DBI::dbGetQuery}
}
\value{
ir_connect() returns an S4 object into the global environment that inherits from DBIConnection. This object is used to communicate with the database engine. The global object will have the same name as the DSN name.
}
\description{
ir_connect() is a wrapper function for DBI::dbConnect that attempts to streamline connections to a University's Oracle databases. Using this function, it is possible for a user to specify only the Data Source Name (DSN).

If a the user id or password for the DSN are not specified, ir_connect() will prompt the user to enter their credentials via the ir_credentials() function.
}
