% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_credentials.R
\name{ir_credentials}
\alias{ir_credentials}
\title{Install database credentials in your \code{.Renviron} file for repeated use}
\usage{
ir_credentials(username, password, install = FALSE, overwrite = FALSE)
}
\arguments{
\item{username}{Your usernamed used for accessing databases formatted in quotes.}

\item{password}{The password used for accessing databases formatted in quotes.}

\item{install}{If TRUE, will install the key in your \code{.Renviron} file
for use in future sessions.  Defaults to FALSE (single session use).}

\item{overwrite}{If TRUE, will overwrite existing credentials that you already have in your \code{.Renviron} file.}
}
\description{
This function adds your reseptive system username and password to your
\code{.Renviron} file so it can be called securely without being stored in
your code. After you have installed these two credentials, they can be
called any time with \code{Sys.getenv("YOUR_USERNAME")} or
\code{Sys.getenv("YOUR_PASSWORD")}. If you do not have an
\code{.Renviron} file, the function will create one for you. If you already
have an \code{.Renviron} file, the function will append the key to your
existing file, while making a backup of your original file for disaster
recovery purposes.
}
