% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ir_pins.R
\name{ir_pins}
\alias{ir_pins}
\title{A wrapper function for pins()}
\usage{
ir_pins(use, board = board_rsconnect(), name = NULL, x = NULL)
}
\arguments{
\item{use}{Type of function desired. Options include 'list' (pin_list), 'read' (pin_read), or 'write' (pin_write).}

\item{board}{Borad to connect to. Default is pins::board_resconnect()}

\item{name}{Pin name to either write or retrieve}

\item{x}{An object (typically a data frame) to pin.}
}
\description{
ir_pins() is a wrapper function for pins::pin_list, pins::pin_read, and pins::pin_get() that attempts to streamline connections to the pins board.
}
