% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iRtools-package.R
\docType{package}
\name{iRtools-package}
\alias{iRtools}
\alias{iRtools-package}
\title{iRtools: A Package for Institutional Research Offices}
\description{
The iRtools package consists of functions both original and found to help Institutional Research officies of all sizes complete work quicker and easier.
}
\author{
\strong{Maintainer}: Bradley Fetes \email{bfetes@syr.edu}

Authors:
\itemize{
  \item Christopher cdunham@syr.edu Dunham
  \item Scott sdavis05@syr.edu Davis
  \item Amy akgardin@syr.edu Gardiner
}

}
\keyword{internal}
