% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lakeDashboard.r
\name{lakeDashboard}
\alias{lakeDashboard}
\title{Run lake dashboard shiny app
Runs lake dashboard shiny app embedded in UT IR tools package.}
\usage{
lakeDashboard()
}
\description{
Run lake dashboard shiny app
Runs lake dashboard shiny app embedded in UT IR tools package.
}
