% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepAttains.R
\name{prepAttains}
\alias{prepAttains}
\title{Prepares rolled up AU-Use and AU-Use-Parameter assessments for ATTAINS upload}
\usage{
prepAttains(current_assessment, trans_wb, pull_attains = TRUE,
  assess_path, param_path, last_reporting_cycle, current_reporting_cycle,
  filesave = TRUE)
}
\arguments{
\item{current_assessment}{A list of assessment output objects to be rolled up to a different spatial resolution.}

\item{trans_wb}{Excel workbook containing translation tables connecting irTools outputs to ATTAINS inputs.}

\item{pull_attains}{Logical. If TRUE (default), function uses readAttains function from the irTools package to pull assessment data from the previous reporting cycle. If FALSE, param_path and assess_path are needed.}

\item{assess_path}{File path to ATTAINS assessments.csv from previous reporting cycle.}

\item{param_path}{File path to ATTAINS parameters.csv from previous reporting cycle.}

\item{last_reporting_cycle}{Year of previous IR reporting cycle.}

\item{current_reporting_cycle}{Year of current IR reporting cycle.}

\item{filesave}{Logical. If TRUE (default), function saves AU-Use-Assessments and Category 5 AU-Use-Parameter Assessments to .csvs in a folder selected by the user.}
}
\value{
Returns a list consisting of two objects: AU-Use-Assessments for all AU's and Category 5 AU-Use-Parameter Assessments.
}
\description{
Prepares rolled up AU-Use and AU-Use-Parameter assessments for ATTAINS upload
}
