% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullAttains.r
\name{pullAttains}
\alias{pullAttains}
\title{Pull data from EPA's ATTAINS database}
\usage{
pullAttains(type = "assessments", ...)
}
\arguments{
\item{type}{ATTAINS data type to pull. One of: "assessmentUnits", "assessments", "actions", or "domains".}

\item{...}{Additional arguments to be passed to ATTAINS web service path.}

\item{stateCode}{Two letter state code (for all pulls where type != "assessments")}

\item{state}{Two letter state code (for pull type == "assessments")}

\item{reportingCycle}{Character string of four digit year of interest (NOTE: EPA working on update that will return the most recent cycle)}
}
\value{
Returns a data.frame of queried ATTAINS data. Note that some results data.frames may contain additional lists of data.frames that can be further flattened.
}
\description{
This function pulls information from EPA ATTAINS database based on submitted arguments via ATTAINS web service. Any ATTAINS web service compatible argument can be submitted to this funciton. Depending on type, at least one argument may be required. See https://link-to-be-obtained.gov for a list of required and compatible arguments.
The function is essentially an ATTAINS specific wrapper for jsonlite::fromJSON. It generates the appropriate web path to connect to ATTAINS web service and converts JSON to R object.
}
\examples{
#Pull South Dakota assessment unit information
SD_AUs=pullAttains(type="assessmentUnits", stateCode="SD")

#Pull South Dakota actions
SD_actions=pullAttains(type="actions", stateCode="SD")
}
