% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readUses.r
\name{readUses}
\alias{readUses}
\title{Read beneficial uses table from R317-2}
\usage{
readUses(file, col_names = c("Name", "1C", "2A", "2B", "3A", "3B", "3C",
  "3D", "3E", "4"))
}
\description{
Note - This is not a fully self-sufficient function. Some substantial manual editing was performed on the file before reading see:
}
