% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readWQPFiles.R
\name{readWQPFiles}
\alias{readWQPFiles}
\title{Read WQP Files}
\usage{
readWQPFiles(narrowresult_file, sites_file, activity_file,
  detquantlim_file, orph_check = TRUE)
}
\arguments{
\item{narrowresult_file}{Full path and filename of narrowresult file queried from WQP to be reviewed (.csv).}

\item{sites_file}{Full path and filename of sites file queried from WQP to be reviewed (.csv).}

\item{activity_file}{Full path and filename of activity file queried from WQP to be reviewed (.csv).}

\item{detquantlim_file}{Full path and filename of detquantlim file queried from WQP to be reviewed (.csv).}

\item{orph_check}{Logical. Specifies whether function should perform orphan check between narrowresult and sites, activity, and detquantlim. If TRUE, produces list object containing orphan rows.}

\item{num_check}{Logical. Specifies whether function should perform checks for non-numeric values in numeric columns (ResultMeasureValue and DetectionQuantitationLimitMeasure.MeasureValue). Opens edit window to make changes to non-numeric values in the data.}
}
\value{
List containing dataframes needed to run irTools package, as well as orphan records and edited non-numeric data tables.
}
\description{
Creates list object containing data frames needed to run irTools. Removes duplicates and creates objects containing orphan records. Identifies non-numeric values in numeric columns and allows user to edit values and save to new object.
}
