% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_arwu.R
\name{scrape_arwu}
\alias{scrape_arwu}
\title{Scrape ARWU table}
\usage{
scrape_arwu(year)
}
\arguments{
\item{year}{An integer. Currently, only years from 2004 onward # 'are compatible due to a change in the table layout at this point.}
}
\value{
A dataframe containing ARWU rankings and scores for the given
year.
}
\description{
Scrapes an ARWU rankings table for a specified year from the following
page: \url{https://www.shanghairanking.com/rankings/arwu/2021} (2021 shown, but
other years can be specified)
}
\details{
Also calculates a "computed score". As ARWU only publishes the exact total
score and rank for the top 100 institutions, we use the ARWU formula to
compute the total score from component scores, and use the computed score
to calculate the exact rank. Note that the computed score is different than the #' the total score due to scaling that occurs to make the score of the top
institution 100.
}
