% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siteValApp.r
\name{siteValApp}
\alias{siteValApp}
\title{Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.}
\usage{
siteValApp()
}
\description{
Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.
}
