% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suppress_repeats.R
\name{suppress_repeats}
\alias{suppress_repeats}
\title{Suppress repeated row labels in summary tables}
\usage{
suppress_repeats(.data, columns = NULL)
}
\arguments{
\item{.data}{a grouped data frame}

\item{columns}{an optional character vector of column names on which to operate}
}
\value{
A tbl, no longer grouped
}
\description{
suppress_repeats() is a function designed to improve the readibility of summary tables. It works best with grouped data frames, just prior to creating your output.
}
