% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_trans.R
\name{term_trans}
\alias{term_trans}
\title{Translate or transform a column with term-related data.}
\usage{
term_trans(col, trans, trans_season = NULL)
}
\arguments{
\item{col}{The column to be translated or transformed. Acceptable term-related data includes term code (of either character or numerical type), term name, and academic year.}

\item{trans}{The desired translation or transformation. Translations include 'term code', 'term name', 'acad year', 'term begin date', and 'term end date'. Translations to term dates correspond to the undergraduate academic calendar. Academic year data can be translated into 'term code' and 'term name' only. Transformations include 'next', 'previous', 'next fall spring', 'previous fall spring', 'next fall', 'previous fall', 'next spring', and 'previous spring'. Academic year data can be transformed into 'next' and 'previous' only.}

\item{trans_season}{Only required when translating academic year data. Specify 'fall', 'spring', or 'summer'.}
}
\value{
Returns a vector of translated or transformed term-related data.
}
\description{
term_trans() takes a dataframe column of term-related data and translates it into different values that represent the same term, or transforms the data into values that represent past or future terms.
}
