% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ir.R
\name{theme_ir}
\alias{theme_ir}
\title{An Office of Institutional Research Theme}
\usage{
theme_ir(base_size = 12, ticks = TRUE, gridlines = FALSE)
}
\arguments{
\item{base_size}{base font size}

\item{ticks}{logical Show axis ticks?}

\item{gridlines}{logical Show major grid lines?}
}
\description{
A clean theme for presenting various visuals, plots, and graphics
}
\note{
The default font family is set to 'Sherman Sans.' theme_ir() attaches extrafont automatically. If Sherman fonts are not avaialble, they will be substituted with generic system fonts.
}
