% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCommentTable.r
\name{updateCommentTable}
\alias{updateCommentTable}
\title{Identify New WQP comment combinations and append them into IR Parameter Translation Workbook for review (sheet name activityCommentTable)}
\usage{
updateCommentTable(data, translation_wb,
  sheetname = "activityCommentTable", startRow = 1, startCol = 1)
}
\arguments{
\item{data}{A WQP results (must include activity, merged objects OK) R-object or the full path and filename to a WQP result file (.csv). If input is a path/file name, datatype must be specified as "filepath".}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{sheetname}{Name of sheet in workbook holding comment table. Defaults to "activityCommentTable".}

\item{startRow}{Row to start reading the activityCommentTable excel sheet from (in case headers have been added). Defaults to 1.}

\item{startCol}{Col to start writing the activityCommentTable excel sheet to (to the right of all formula based columns). Defaults to 1.}
}
\value{
Appends any new combinations in WQP data to translation_wb for review. This updates the input translation_wb with those new rows with system date in the "DateAdded" column.
}
\description{
This function identifies new ResultLaboratoryCommentCode/ResultLaboratoryCommentText/ActivityCommentText/ResultCommentText combinations in WQP data for which an IR_FLAG has not yet been defined and appends them into the translation workbook for review.
}
