% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateDetCondLimTables.r
\name{updateDetCondLimTables}
\alias{updateDetCondLimTables}
\title{Update detection condition and limit type tables}
\usage{
updateDetCondLimTables(
  results,
  detquantlim,
  translation_wb,
  detConditionTable_sheetname = "detConditionTable",
  detConditionTable_startRow = 2,
  detConditionTable_startCol = 1,
  detLimitTypeTable_sheetname = "detLimitTypeTable",
  detLimitTypeTable_startRow = 2,
  detLimitTypeTable_startCol = 1
)
}
\arguments{
\item{results}{A WQP results (must include narrow result, merged, wide objects OK) R-object name.}

\item{detquantlim}{A WQP detection/quantitation limit file R-object. Should be matching complement to WQP results input. New values derived from those that successfully merge to results input.}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{detConditionTable_sheetname}{Name of sheet in workbook holding result detection condition names table. Defaults to "detConditionTable".}

\item{detConditionTable_startRow}{Row to start reading the detConditionTable excel sheet from (in case headers have been added). Defaults to 2.}

\item{detConditionTable_startCol}{Col to start writing the detConditionTable excel sheet to (to the right of all formula based columns). Defaults to 1.}

\item{detLimitTypeTable_sheetname}{Name of sheet in workbook holding detection limit type names and ranked prioritizations table. Defaults to "detLimitTypeTable".}

\item{detLimitTypeTable_startRow}{Row to start reading the detLimitTypeTable excel sheet from (in case headers have been added). Defaults to 2.}

\item{detLimitTypeTable_startCol}{Col to start writing the detLimitTypeTable excel sheet to (to the right of all formula based columns). Defaults to 1.}
}
\value{
Appends any new values in WQP data to translation_wb for review. This updates the input translation_wb with those new rows.
}
\description{
This function identifies new ResultDetectionConditionText & DetectionQuantitationLimitTypeName values in WQP data for which are not in IR translation tables and appends them into the translation workbook for review.
New ResultDetectionConditionText & DetectionQuantitationLimitTypeName values generated by this function are derived from those in detquantlim that match (via merge) records in results input (this allows subsetting of these lists by other previously generated data screens if desired).
}
