% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParamTransTable.r
\name{updateParamTrans}
\alias{updateParamTrans}
\title{Update paramTransTable, paramFractionGroup, & WQPParamCASID tables in translation workbook}
\usage{
updateParamTrans(
  data,
  translation_wb,
  standards_wb,
  paramTransTable_sheetname = "paramTransTable",
  paramTransTable_startRow = 1,
  paramTransTable_startCol = 1,
  CASLookup_sheetname = "CASLookup",
  CASLookup_startRow = 2,
  CASLookup_startCol = 1,
  paramFractionGroup_sheetname = "paramFractionGroup",
  paramFractionGroup_startRow = 2,
  paramFractionGroup_startCol = 2
)
}
\arguments{
\item{data}{A WQP results (merged or narrow result) R-object}

\item{translation_wb}{Full path and filename for IR translation workbook}

\item{standards_wb}{Full path and filename for IR standards workbook}

\item{paramTransTable_sheetname}{Name of sheet in workbook holding parameter translation table. Defaults to "paramTransTable".}

\item{paramTransTable_startRow}{Row to start reading the paramTransTable excel sheet from (in case additional headers have been added). Defaults to 1.}

\item{paramTransTable_startCol}{Col to start writing the paramTransTable excel sheet to (to the right of all formula based columns). Defaults to 1.}

\item{CASLookup_sheetname}{Name of sheet in workbook holding CAS lookup table from WQP. Defaults to "CASLookup".}

\item{CASLookup_startRow}{Row to start reading the CASLookup excel sheet from (in case additional headers have been added). Defaults to 2.}

\item{CASLookup_startCol}{Col to start writing the CASLookup excel sheet to (to the right of all formula based columns). Defaults to 1.}

\item{paramFractionGroup_sheetname}{Name of sheet in workbook holding the WQP fraction names. Defaults to "paramFractionGroup".}

\item{paramFractionGroup_startRow}{Row to start reading the paramFractionGroup excel sheet from (in case additional headers have been added). Defaults to 2.}

\item{paramFractionGroup_startCol}{Col to start writing the paramFractionGroup excel sheet to (to the right of all formula based columns, note col w/ header 'spacer' in table). Defaults to 2.}
}
\value{
Appends any new parameter/fraction/unit combinations in WQP data to translation_wb. This updates the input translation_wb with those new rows with system date in the "DateAdded" column.
}
\description{
This function identifies parameter/fraction/unit/method combinations in WQP data for which an IR translation has not yet been defined and appends them to the translation table for review.
}
