% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateTable.R
\name{updateTable}
\alias{updateTable}
\title{Update translation workbook sheets with new combinations found in the WQP dataset.}
\usage{
updateTable(data, trans_wb, sheet, cols, params = FALSE, standards_wb)
}
\arguments{
\item{data}{A merged WQP result object. Must include both narrowresult & activity OR detquantlim files. May also be a post-fillMaskedValues() results object. Note: re-application of edited domain tables to an already screened dataset is not advised, and changes to the domain table likely will not be reflected in a re-screened dataset due to merging conventions.}

\item{trans_wb}{Full path and filename for Excel workbook containing screen table to be applied}

\item{sheet}{Name of sheet in workbook holding desired screening decisions}

\item{cols}{A vector of column names to be screened for new combinations of values to be placed in the translation workbook.}

\item{params}{Logical. If true, will screen WQ parameters for CAS ID and match to CAS IDs in the standards workbook.}

\item{standards_wb}{Full path and filename for Excel workbook containing WQ standards to be used to assess WQP data.}
}
\value{
Message reporting the number of new combinations found in the data that were added to the workbook sheet. Workbook sheet is updated in translation workbook, with blank IR_FLAG column for new combinations in need of review.
}
\description{
Joins translation workbook sheets (activityMediaNameTable, labNameActivityTable, detConditionTable, detLimitTypeTable, paramTransTable, paramFractionTable, unitConvTable) to WQP data to apply decisions from input files to data.
}
