% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQPdataqaqc.r
\name{WQPdataqaqc}
\alias{WQPdataqaqc}
\title{WQP Data QA/QC Screen}
\usage{
WQPdataqaqc(outfile_path, narrowresult_file, sites_file, activity_file,
  detquantlim_file, dupcheck = c("narrowresult", "sites", "activity",
  "detquantlim"))
}
\arguments{
\item{outfile_path}{Path to folder where user wishes to deposit orphan check .csv's}

\item{narrowresult_file}{Full path and filename of narrowresult file queried from WQP to be reviewed (.csv).}

\item{sites_file}{Full path and filename of sites file queried from WQP to be reviewed (.csv).}

\item{activity_file}{Full path and filename of activity file queried from WQP to be reviewed (.csv).}

\item{detquantlim_file}{Full path and filename of detquantlim file queried from WQP to be reviewed (.csv).}

\item{dupcheck}{Vector of file names to be run through duplicate check. Allows user to save time if they already know whether some files do/do not contain duplicates. Defaults to c("narrowresult", "sites", "activity", "detquantlim)}
}
\value{
Exports .csv files for all orphan checks in specified output path.
}
\description{
Performs duplicate and orphan checks on data downloaded from WQP.
Alerts user to potential issues with running data through subsequent irTools functions.
}
