% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyScreenTable.r
\name{applyScreenTable}
\alias{applyScreenTable}
\title{Apply screens to WQP data by selected table}
\usage{
applyScreenTable(data, datatype = "object", translation_wb, sheetname,
  flag_col_name, com_col_name, startRow = 1)
}
\arguments{
\item{data}{Either a merged result object (narrowresult<-activity<-quantdetlim) or the full path and filename to a merged result .csv file. If data input is a path/file name, restype must be specified as "filepath".}

\item{datatype}{Data input type (one of "object" or "filepath"). Defaults to "object". Specify "filepath" to input a path and file name for the results function argument.}

\item{translation_wb}{Full path and filename for IR translation workbook}

\item{sheetname}{Name of sheet in workbook holding desired screening decisions}

\item{flag_col_name}{Name to rename IR_FLAG column to.}

\item{com_col_name}{Name to rename IR_COMMENT column to.}

\item{startRow}{Row to start reading excel sheet from (in case additional headers have been added). Defaults to 1.}
}
\value{
A data.frame object of WQP data with merged columns from input screening tables.
}
\description{
Joins activity review inputs (labNameActivityTable, activityMediaNameTable, & activityCommentTable) to WQP data to apply decisions from input files to data.
}
