% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessHFDO.R
\name{assessHFDO}
\alias{assessHFDO}
\title{Run HFDO assessments}
\usage{
assessHFDO(data, consecday = 39)
}
\arguments{
\item{data}{HFDO data--for the time being is a test dataset built by EH using "hfdo_prep.R".}

\item{consecday}{Numeric. Minimum number of consecutive days needed to perform an assessment.}
}
\value{
List object containing data used for minimum, 7-day, and 30-day assessments, full list of assessments by type, and a rollup to site-use assessments.
}
\description{
Performs high frequency dissolved oxygen assessments per IR assessment methods. This includes checking for data sufficiency, calculating daily minima and averages, and 7-day/30-day moving averages.
}
