% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessLakeProfiles.r
\name{assessLakeProfiles}
\alias{assessLakeProfiles}
\title{Run lake profile assessments}
\usage{
assessLakeProfiles(data)
}
\arguments{
\item{data}{}
}
\value{
Returns dataframe with assessment categories for each AU/BenUse/R3172ParameterName.
}
\description{
Description description description
}
