% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignCriteria.r
\name{assignCriteria}
\alias{assignCriteria}
\title{Assign numeric criteria to WQP data}
\usage{
assignCriteria(data, crit_wb, crit_sheetname, ss_sheetname, crit_startRow = 1,
  ss_startRow = 1)
}
\arguments{
\item{data}{A merged, parameter translated, spatially referenced WQP result object. Must be post parameter translation step.}

\item{crit_wb}{Full path and filename for workbook containing criteria.}

\item{crit_sheetname}{Name of sheet in workbook holding criteria to be assigned.}

\item{ss_sheetname}{Name of sheet in workbook holding site-specific criteria to be assigned. Should typically contain parameters corresponding to those in criteria table.}

\item{crit_startRow}{Row to start reading criteria table excel sheet from (in case headers have been added). Defaults to 1.}

\item{ss_startRow}{Row to start reading site specific criteria table excel sheet from (in case headers have been added). Defaults to 1.}
}
\value{
A flattened & expanded sample x use x parameter x criterion data.frame.
}
\description{
Assigns general & site-specific numeric WQ criteria to WQP data. Does not calculate criteria w/ correction factors. See calcCriteria (function in development).
}
