% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoValidateWQPsites.r
\name{autoValidateWQPsites}
\alias{autoValidateWQPsites}
\title{Auto-validate USEPA WQP stations}
\usage{
autoValidateWQPsites(sites_object, master_site_file, correct_longitude = FALSE)
}
\arguments{
\item{sites_object}{Sites object queried from WQP to be reviewed.}

\item{master_site_file}{Full path and filename of master site list as generated by autoValidateWQPsites function and manual site review application. Also contains a sheet of waterbody types and their associated IR_FLAGs (.xlsx).}
}
\value{
Exports a new, undated master site list to the location & filename provided by the user.
}
\description{
Performs auto-validation on previously queried WQP stations combined with the existing master site list.
Auto-validates any new sites (i.e. those not currently in the master site list) and appends them into existing master site list flagged for acceptance, rejection, or review.
Checks for any new site types in new data. A warning message and a list of new site types is printed if new site types are encountered.
Also re-auto-validates any sites in the master site list that have previously only undergone auto-validation (to account for any changes in auto validation process) and checks existing master site list for changes in AUs, selected site types, or property boundaries.
}
