% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepExportData.R
\name{composeExport}
\alias{composeExport}
\title{Create dataset for secondary reviewer export from assessment dashboard}
\usage{
composeExport(prepped_data)
}
\arguments{
\item{prepped_data}{A list of objects produced by dataPrep function, including toxics, conventionals, and ecoli dataframes.}
}
\value{
A list composed of three dataframes: site-date-use-param records linked with aggregated daily values (if applicable), calculations, and exceedances, site-use-param summaries including sample and exceedance counts, and prepped E.coli data.
}
\description{
Pulls columns of interest from prepped data and calculates basic exceedance summaries for site-use-parameter assessments.
}
