% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataPrep.r
\name{dataPrep}
\alias{dataPrep}
\title{Final data prep step before assessments}
\usage{
dataPrep(data, translation_wb, unit_sheetname = "unitConvTable",
  startRow = 1, split_agg_tds = TRUE)
}
\arguments{
\item{data}{A merged, translated, and numeric criteria assigned WQP results R-object. Target units for conversions are defined by units associated with assigned numeric critera.}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{unit_sheetname}{Name of sheet in workbook holding IR unit conversion table. Defaults to "unitConvTable".}

\item{startRow}{Row to start reading the unit conversion table excel sheet from (in case headers have been added). Defaults to 1.}

\item{split_agg_tds}{Logical. If TRUE (default) split off TDS records w/ function assigned in AsmntAggFun into separate output. If FALSE, these records are passed through to conventionals output.}
}
\value{
A list of objects ready for assessments.
}
\description{
Performs unit conversions & daily data aggregations. Also checks native vs. target activities and fractions, dissolved vs. total value checks, river/stream depth & flow checks, and generates value based data flags.
}
