% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepWQPData.R
\name{prepWQPData}
\alias{prepWQPData}
\title{Prep WQP Data}
\usage{
prepWQPData(irdata, check = TRUE)
}
\arguments{
\item{nr}{A narrow result dataframe from EPA's WQP}

\item{activity}{An activity result dataframe from EPA's WQP}

\item{detquantlim}{A detection quantitation limits dataframe from EPA's WQP}

\item{sites}{A sites dataframe from EPA's WQP}
}
\value{
A list consisting of nr, activity, detquantlim, sites, filtered sites, and merged_results (nr combined with activity) dataframes with special characters converted to NA for the ResultMeasureValue column. Saves an rdata file named "irdata_wqp_raw.Rdata" containing the six data WQP objects.
}
\description{
This function reads dataframes downloaded from EPA's WQP, removes duplicates, and converts result values with non-numeric characters to NA (with the exception of those with <|>|,).
Replaces hard coding in bookdown and readWQPFiles function.
}
