% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullWQP.r
\name{pullWQP}
\alias{pullWQP}
\title{Pull water quality, station, and other data for Utah from USEPA WQP}
\usage{
pullWQP(outfile_path, StartDate, EndDate, retrieve = c("narrowresult",
  "activity", "sites", "detquantlim"), retry = FALSE)
}
\arguments{
\item{outfile_path}{Path for file outputs.}

\item{StartDate}{Query start date. "MM-DD-YYYY" format.}

\item{EndDate}{Query end date. "MM-DD-YYYY" format.}

\item{retrieve}{Vector of data type names to retrieve from WQP. One or more of: "result","narrowresult","activity","activitymetric","sites","detquantlim". Defaults to query all.}
}
\value{
Exports .csv files for all selected data types during selected date period in specified output path.
}
\description{
Pulls data from EPA's Water Quality Portal (WQP) and exports data files (as .csv) to outfile_path folder.
Prepared by C Shope 1/25/18, Adapted from Roop Guha-NJDEP 1/3/18.
Updated, re-formated to function, & packaged by J VanderLaan.
}
