% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollUp.R
\name{rollUp}
\alias{rollUp}
\title{Summarize assessments to different spatial resolutions}
\usage{
rollUp(data, group_vars = c("ASSESS_ID", "BeneficialUse",
  "R3172ParameterName"), cat_var = "IR_Cat", expand_uses = TRUE,
  print = TRUE)
}
\arguments{
\item{data}{A list of assessment output objects to be rolled up to a different spatial resolution.}

\item{group_vars}{Vector of column names on which to group data for assessment rollups. Defaults to aggregate by ASSESS_ID, BeneficialUse, and R3172ParameterName.}

\item{cat_var}{Name of assessment category column.}

\item{expand_uses}{Logical. If TRUE (default), uses are expanded in the output to include all uses associated with group_vars, including unassessed groups which are marked as 'NA' in output dataframe column AssessCat.  If FALSE, only assessed groups are included in the output.}

\item{print}{Logical. If TRUE (default) print summary table of applicable assessment categories & parameters.}
}
\value{
Returns dataframe with assessment categories for each AU/BenUse/R3172ParameterName.
}
\description{
Summarizes and aggregates assessments at one spatial resolution to another spatial resolution. For example, a collection of assessments conducted at the site/use/parameter level can be rolled up to the AU/use/parameter level or to the AU/use/level.
}
