% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSiteValApp.r
\name{runSiteValApp}
\alias{runSiteValApp}
\title{Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.}
\usage{
runSiteValApp()
}
\description{
Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.
}
