% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSiteValApp.r
\name{runSiteValApp}
\alias{runSiteValApp}
\title{Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.}
\usage{
runSiteValApp(master_site_file, polygon_path, edit_log_path, reasons_flat_file)
}
\arguments{
\item{master_site_file}{Full path and filename of master site list as generated by autoValidateWQPsites function and manual site review application (.csv).}

\item{polygon_path}{Full path to folder containing BU, AU, and state of UT polygons.}

\item{edit_log_path}{Full path to folder containing site edit log.}

\item{reasons_flat_file}{Full patch and filename of flat site review/reject reasons exported by autoValidateWQPsites function.}
}
\description{
Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.
}
