% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSiteValApp.r
\name{runSiteValApp}
\alias{runSiteValApp}
\title{Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.}
\usage{
runSiteValApp(master_site_file, polygon_path, edit_log_path)
}
\arguments{
\item{master_site_file}{Full path and filename of master site list as generated by autoValidateWQPsites function and manual site review application (.csv).}
}
\description{
Run site validation shiny app
Runs site validation shiny app embedded in UT IR tools package.
}
