% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateLabActMediaTables.r
\name{updateLabActMediaTables}
\alias{updateLabActMediaTables}
\title{Update labNameActivityTable & activityMediaNameTable}
\usage{
updateLabActMediaTables(data, translation_wb,
  labNameActivityTable_sheetname = "labNameActivityTable",
  labNameActivityTable_startCol = 1, labNameActivityTable_startRow = 1,
  activityMediaNameTable_sheetname = "activityMediaNameTable",
  activityMediaNameTable_startCol = 1,
  activityMediaNameTable_startRow = 1)
}
\arguments{
\item{data}{A WQP results (must include activity, merged objects OK) R-object}

\item{translation_wb}{Full path and filename for IR translation workbook (.xlsx).}

\item{labNameActivityTable_sheetname}{Name of sheet in workbook holding lab name / activity table. Defaults to "labNameActivityTable".}

\item{labNameActivityTable_startCol}{Col to start writing the labNameActivityTable excel sheet to (to the right of all formula based columns). Defaults to 3.}

\item{labNameActivityTable_startRow}{Row to start reading the labNameActivityTable excel sheet from (in case headers have been added). Defaults to 1.}

\item{activityMediaNameTable_sheetname}{Name of sheet in workbook holding activity / media/submedia name table. Defaults to "activityMediaNameTable".}

\item{activityMediaNameTable_startCol}{Col to start writing the activityMediaNameTable excel sheet to (to the right of all formula based columns). Defaults to 3.}

\item{activityMediaNameTable_startRow}{Row to start reading the activityMediaNameTable excel sheet from (in case headers have been added). Defaults to 1.}
}
\value{
Appends any new combinations in WQP data to translation_wb for review. This updates the input translation_wb with those new rows with system date in the "DateAdded" column.
}
\description{
This function identifies new LaboratoryName/ActivityTypeCode combinations & ActivityMediaName/ActivityMediaSubdivisionName combinations in WQP data for which an IR_FLAG has not yet been defined and appends them into the translation workbook for review.
}
