% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateParamTransTable.r
\name{updateParamTrans}
\alias{updateParamTrans}
\title{Identify New WQP Parameters/Fractions/Units/Methods and Append to IR Parameter Translation Table}
\usage{
updateParamTrans(data, detquantlim = detquantlim, translation_wb,
  paramTransTable_sheetname = "paramTransTable",
  paramTransTable_startRow = 4, paramTransTable_startCol = 16,
  WQPParamCASID_sheetname = "WQPParamCASID", WQPParamCASID_startRow = 4,
  WQPParamCASID_startCol = 1,
  paramFractionGroup_sheetname = "paramFractionGroup",
  paramFractionGroup_startRow = 3, paramFractionGroup_startCol = 3,
  fun_cols = c("CASLinked", "R3172ParamAssessmentType", "CASLinked",
  "ParamAssessmentType", "ParamMeasureType", "AssessmentType",
  "R3172ParameterStatus", "R3172ParameterName", "KeytoPrevIRParamTble",
  "FractionGroup", "ToLabHoldingTime", "ToExtractHoldingTime",
  "AfterExtractHoldingTime", "HoldingTimeFrozen", "TotalHoldingTime",
  "HoldingTimeUnit", "Source"))
}
\arguments{
\item{data}{A WQP results (merged or narrow result) R-object}

\item{detquantlim}{WQP detection/quantitation limit object}

\item{translation_wb}{Full path and filename for IR translation workbook}

\item{paramTransTable_sheetname}{Name of sheet in workbook holding parameter translation table. Defaults to "paramTransTable".}

\item{paramTransTable_startRow}{Row to start reading the paramTransTable excel sheet from (in case additional headers have been added). Defaults to 4.}

\item{paramTransTable_startCol}{Col to start writing the paramTransTable excel sheet to (to the right of all formula based columns). Defaults to 16.}

\item{WQPParamCASID_sheetname}{Name of sheet in workbook holding parameter/CAS ID table. Defaults to "WQPParamCASID".}

\item{WQPParamCASID_startRow}{Row to start reading the WQPParamCASID excel sheet from (in case additional headers have been added). Defaults to 3.}

\item{WQPParamCASID_startCol}{Col to start writing the WQPParamCASID excel sheet to (to the right of all formula based columns). Defaults to 1.}

\item{paramFractionGroup_sheetname}{Name of sheet in workbook holding the WQP fraction names. Defaults to "paramFractionGroup".}

\item{paramFractionGroup_startRow}{Row to start reading the paramFractionGroup excel sheet from (in case additional headers have been added). Defaults to 3.}

\item{paramFractionGroup_startCol}{Col to start writing the paramFractionGroup excel sheet to (to the right of all formula based columns). Defaults to 3.}

\item{fun_cols}{Vector of names of columns in any table that contain functions. These are excluded from reading and writing by this function.}
}
\value{
Appends any new parameter/fraction/unit combinations in WQP data to translation_wb. This updates the input translation_wb with those new rows with system date in the "DateAdded" column.
}
\description{
This function identifies parameter/fraction/unit/method combinations in WQP data for which an IR translation has not yet been defined and appends them to the translation table for review.
}
